pyproject.toml
[tool.poetry]
name = "epic_reminder"
version = "0.1.0"
description = "Reminds when you should do stuff in Epic RPG"
authors = ["John-Paul Jorissen <jjorissen52@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
"discord.py" = "^1.5.1"
Django = "^3.1.3"
python-dotenv = "^0.15.0"
execution-pipeline = "^0.4.0"
pytz = "^2020.4"
aiologger = {extras = ["aiofiles"], version = "^0.6.0"}
django-extensions = "^3.1.1"

[tool.poetry.dev-dependencies]
pre-commit = "^2.8.2"
black = "^20.8b1"
pip = "^20.2.4"
install = "^1.3.4"
fire = "^0.3.1"
watchgod = "^0.7"
ipython = "^7.22.0"

[tool.black]
line-length=120
exclude = """
/(
    ^migrations/
)/
"""

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
